+++ video_ingest_tool/__init__.py
@@ -1,6 +1,7 @@
 """
 AI-Powered Video Ingest & Catalog Tool
-A tool for automating the analysis, categorization, and retrieval of video content.
+
+A tool for video content discovery, metadata extraction, and cataloging.
 """
 
 __version__ = "0.1.0"
